+++ admin/app/AppLoader.scala
@@ -40,7 +40,7 @@ trait AdminServices extends I18nComponents  {
   lazy val emailService = wire[EmailService]
   lazy val fastlyStatisticService = wire[FastlyStatisticService]
   lazy val fastlyCloudwatchLoadJob = wire[FastlyCloudwatchLoadJob]
-  lazy val redirects = wire[Redirects]
+  lazy val redirects = wire[RedirectService]
   lazy val r2PagePressJob = wire[R2PagePressJob]
   lazy val videoEncodingsJob = wire[VideoEncodingsJob]
   lazy val matchDayRecorder = wire[MatchDayRecorder]
