+++ archive/app/AppLoader.scala
@@ -13,7 +13,7 @@ import play.api.http.{HttpErrorHandler, HttpRequestHandler}
 import play.api.mvc.EssentialFilter
 import play.api.routing.Router
 import play.api._
-import services.{ArchiveMetrics, Redirects}
+import services.{ArchiveMetrics, RedirectService}
 import router.Routes
 
 class AppLoader extends FrontendApplicationLoader {
@@ -22,7 +22,7 @@ class AppLoader extends FrontendApplicationLoader {
 
 trait AppComponents extends FrontendComponents {
 
-  lazy val redirects = wire[Redirects]
+  lazy val redirects = wire[RedirectService]
 
   lazy val healthCheck = wire[HealthCheck]
   lazy val archiveController = wire[ArchiveController]
